Remove mapping::Submaps.
